+++ derive_builder_core/src/builder.rs
@@ -72,7 +72,7 @@ impl<'a> ToTokens for Builder<'a> {
             let builder_fields = &self.fields;
             let functions = &self.functions;
             let builder_doc_comment = &self.doc_comment;
-            let deprecation_notes = &self.deprecation_notes;
+            let deprecation_notes = &self.deprecation_notes.as_item();
 
             debug!("ty_generics={:?}, where_clause={:?}, impl_generics={:?}", ty_generics, where_clause, impl_generics);
 
@@ -86,8 +86,8 @@ impl<'a> ToTokens for Builder<'a> {
                 #[allow(dead_code)]
                 impl #impl_generics #builder_ident #ty_generics #where_clause {
                     #(#functions)*
+                    #deprecation_notes
                 }
-                #deprecation_notes
             ));
         } else {
             trace!("Skipping builder `{}`.", self.ident);