+++ derive_builder_core/CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- struct default
+
 ## 0.1 - 2017-03-25
 ### Added
 - helper crate `derive_builder_core`: